Moved table row into its own folder
