+++ react.app/src/components/pages/library_items/LibraryItemsTable/LibraryItemsTableRow/LibraryItemCheckinDialog.tsx
@@ -5,49 +5,40 @@ import {
   DialogContent,
   DialogContentText,
   DialogTitle,
-  TextField,
 } from '@mui/material';
 import React from 'react';
-import { LibraryItem } from '../../../../types';
-import { ItemInformationBox } from '../../../sub_components/ItemInformationBox';
+import { LibraryItem } from '../../../../../types';
+import { ItemInformationBox } from '../../../../sub_components/ItemInformationBox';
 
-interface LibraryItemCheckoutDialogProps {
+interface LibraryItemCheckinDialogProps {
   open: boolean;
   handleClose: () => void;
   item: LibraryItem;
 }
 
-export default function LibraryItemCheckoutDialog(
-  props: LibraryItemCheckoutDialogProps
+export default function LibraryItemCheckinDialog(
+  props: LibraryItemCheckinDialogProps
 ) {
   const { open, handleClose, item } = props;
   return (
     <Dialog
       open={open}
       onClose={handleClose}
-      aria-labelledby="checkout-item-dialog"
+      aria-labelledby="checkin-item-dialog"
     >
-      <DialogTitle id="checkout-item-dialog">Checkout Item</DialogTitle>
+      <DialogTitle id="checkin-item-dialog">Checkin Item</DialogTitle>
       <DialogContent>
         <DialogContentText>
-          Are you sure you want to checkout this item?
+          Are you sure you want to checkin this item?
         </DialogContentText>
         <ItemInformationBox item={item} />
-        <TextField
-          autoFocus
-          margin="dense"
-          id="checkout-item-borrower"
-          label="Borrower's name"
-          fullWidth
-          variant="standard"
-        />
       </DialogContent>
       <DialogActions>
         <Button autoFocus onClick={handleClose} color="primary">
           Cancel
         </Button>
         <Button onClick={handleClose} variant="contained" color="success">
-          Checkout
+          Checkin
         </Button>
       </DialogActions>
     </Dialog>
