+++ react.app/src/components/pages/library_items/LibraryItemsTable/LibraryItemsTableRow/LibraryItemCheckoutDialog.tsx
@@ -5,40 +5,49 @@ import {
   DialogContent,
   DialogContentText,
   DialogTitle,
+  TextField,
 } from '@mui/material';
 import React from 'react';
-import { LibraryItem } from '../../../../types';
-import { ItemInformationBox } from '../../../sub_components/ItemInformationBox';
+import { LibraryItem } from '../../../../../types';
+import { ItemInformationBox } from '../../../../sub_components/ItemInformationBox';
 
-interface LibraryItemCheckinDialogProps {
+interface LibraryItemCheckoutDialogProps {
   open: boolean;
   handleClose: () => void;
   item: LibraryItem;
 }
 
-export default function LibraryItemCheckinDialog(
-  props: LibraryItemCheckinDialogProps
+export default function LibraryItemCheckoutDialog(
+  props: LibraryItemCheckoutDialogProps
 ) {
   const { open, handleClose, item } = props;
   return (
     <Dialog
       open={open}
       onClose={handleClose}
-      aria-labelledby="checkin-item-dialog"
+      aria-labelledby="checkout-item-dialog"
     >
-      <DialogTitle id="checkin-item-dialog">Checkin Item</DialogTitle>
+      <DialogTitle id="checkout-item-dialog">Checkout Item</DialogTitle>
       <DialogContent>
         <DialogContentText>
-          Are you sure you want to checkin this item?
+          Are you sure you want to checkout this item?
         </DialogContentText>
         <ItemInformationBox item={item} />
+        <TextField
+          autoFocus
+          margin="dense"
+          id="checkout-item-borrower"
+          label="Borrower's name"
+          fullWidth
+          variant="standard"
+        />
       </DialogContent>
       <DialogActions>
         <Button autoFocus onClick={handleClose} color="primary">
           Cancel
         </Button>
         <Button onClick={handleClose} variant="contained" color="success">
-          Checkin
+          Checkout
         </Button>
       </DialogActions>
     </Dialog>
