+++ react.app/src/components/sub_components/ItemInformationBox.tsx
@@ -1,4 +1,4 @@
-import { Stack, Chip } from '@mui/material';
+import { Stack, Chip, Paper } from '@mui/material';
 import React from 'react';
 import { LibraryItem } from '../../types';
 
@@ -9,11 +9,13 @@ interface ItemInformationBoxProps {
 export function ItemInformationBox(props: ItemInformationBoxProps) {
   const { item } = props;
   return (
-    <Stack direction="row" spacing={2} justifyContent="center" mt={2}>
-      <Chip label={item.id} color="info"/>
-      <Chip label={item.title} color="info"/>
-      <Chip label={item.type} color="info"/>
-      <Chip label={item.isBorrowable ? 'Available' : 'Unavailable'} color="info"/>
-    </Stack>
+    <Paper elevation={1} sx={{ p: 1, pb: 2 }}>
+      <Stack direction="row" spacing={2} justifyContent="center" mt={2}>
+        <Chip label={item.id} color="info"/>
+        <Chip label={item.title} color="info"/>
+        <Chip label={item.type} color="info"/>
+        <Chip label={item.isBorrowable ? 'Available' : 'Unavailable'} color="info"/>
+      </Stack>
+    </Paper>
   );
 }
